[CI/Build] make it possible to build with a free-threaded interpreter
For Python 3.13/3.14, free-threaded Python does not support using the
Limited C API and the Stable ABI. The purpose of this PR is to ensure
that vLLM can be built from source under a free-threaded interpreter. It
doesn't change anything for default (with-GIL) Python interpreters.

This is an identical change to that in vllm#29242, see that PR for more
context.

Signed-off-by: Ralf Gommers <ralf.gommers@gmail.com>